+++ xui-rp-install-server.sh
@@ -923,28 +923,43 @@ ssh_setup() {
     exec > /dev/tty 2>&1
     msg_inf "Настройка ssh"
     msg_inf "Сгенерируйте ключ для своей ОС (ssh-keygen)"
-    echo    
+    echo
     msg_inf "В windows нужно установить пакет openSSH, и ввести команду в POWERSHELL (предлагаю изучить как генерировать ключ в интернете)"
     msg_inf "Если у вас linux, то вы сами все умеете С:"
     echo
     msg_tilda "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -"
     echo
-    echo -n "Команда для Windows: " && msg_out "type \$env:USERPROFILE\.ssh\id_rsa.pub | ssh -p 22 ${username}@${IP4} \"cat >> ~/.ssh/authorized_keys\""    
+    echo -n "Команда для Windows: " && msg_out "type \$env:USERPROFILE\.ssh\id_rsa.pub | ssh -p 22 ${username}@${IP4} \"cat >> ~/.ssh/authorized_keys\""
     echo -n "Команда для Linux: " && msg_out "ssh-copy-id -p 22 ${username}@${IP4}"
     echo
     msg_tilda "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -"
     echo
-    msg_inf "Настроить ssh (шаг не обязательный)? [y/N]"
-    answer_input
+    while true; do
+        msg_inf "Настроить ssh (шаг не обязательный)? [y/N]"
+        answer_input
 
-    if [[ $? -eq 0 ]]; then
-        sed -i -e "s/#PermitRootLogin/PermitRootLogin/g" -e "s/PermitRootLogin yes/PermitRootLogin prohibit-password/g" /etc/ssh/sshd_config
-        sed -i -e "s/#PubkeyAuthentication/PubkeyAuthentication/g" -e "s/PubkeyAuthentication no/PubkeyAuthentication yes/g" /etc/ssh/sshd_config
-        sed -i -e "s/#PasswordAuthentication/PasswordAuthentication/g" -e "s/PasswordAuthentication yes/PasswordAuthentication no/g" /etc/ssh/sshd_config
-        sed -i -e "s/#PermitEmptyPasswords/PermitEmptyPasswords/g" -e "s/PermitEmptyPasswords yes/PermitEmptyPasswords no/g" /etc/ssh/sshd_config
+        if [[ $? -eq 0 ]]; then
 
-            cat > /etc/motd <<EOF
-        
+            if [[ ! -s /home/${username}/.ssh/id_rsa.pub && ! -s /root/.ssh/id_rsa.pub ]]; then
+                msg_err "Ошибка: Ключи не найдены в файле /home/${username}/.ssh/id_rsa.pub или /root/.ssh/id_rsa.pub"
+                msg_err "Cоздайте ключи и добавьте их на сервер, прежде чем продолжить"
+                msg_inf "Попробуйте снова"
+                echo
+            else
+                # Если ключи найдены, продолжаем настройку SSH
+                sed -i -e "
+                    s/#PermitRootLogin/PermitRootLogin/g;
+                    s/PermitRootLogin yes/PermitRootLogin prohibit-password/g;
+                    s/#PubkeyAuthentication/PubkeyAuthentication/g;
+                    s/PubkeyAuthentication no/PubkeyAuthentication yes/g;
+                    s/#PasswordAuthentication/PasswordAuthentication/g;
+                    s/PasswordAuthentication yes/PasswordAuthentication no/g;
+                    s/#PermitEmptyPasswords/PermitEmptyPasswords/g;
+                    s/PermitEmptyPasswords yes/PermitEmptyPasswords no/g;
+                " /etc/ssh/sshd_config
+                
+                cat > /etc/motd <<EOF
+
 ################################################################################
                          WARNING: AUTHORIZED ACCESS ONLY
 ################################################################################
@@ -979,9 +994,16 @@ to the fullest extent of the law.
 
 
 EOF
-        systemctl restart ssh.service
-        echo "Настройка SSH завершена."
-    fi
+                systemctl restart ssh.service
+                echo "Настройка SSH завершена."
+                break
+            fi
+        else
+            echo "Настройка SSH пропущена."
+            break
+        fi
+    done
+
     echo
     msg_tilda "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -"
     echo
